Update patch set 1
Patch Set 1:

@Geora, without this change, we won't get the info log on line 97 of ContextAggregatorProcessor.java if there is any exceptions from RestRequest.java.  It would just show an error log "For input string: "" ".  

For example, if someone didn't update the configuration for aai context builder, when the context aggregator tries to contact the aai context builder, it throws an exception on line 65 (the current file) of RestRequest.java which skips the rest of the getModelData() and logs an error on line 119 (the current file) of ContextAggregatorProcessor.java.  The error message is so generic that it's hard to identify where was the problem.

Does it make sense?

Patch-set: 1